requirements.txt
flask
gunicorn
Pillow
opencv-python-headless
